Refactor towards more modular approach
A major goal of Grovetender is to focus on what you
want to achieve, and not how. As such, it should work
more like a constructor. You don't create and style a
button, instead you create a panel, and give it a
behavior of a button. You add a text and/or an icon
child to it.

The app then takes care of converting it into a native
button for the target engine, or a more custom
implementation where native types are lacking. This
also means that any widget created with Grovetender
can exhibit behaviors and states commonly applicable
to UI elements. Anything can have a focused style, or
a hover style, or a disabled style.