Introduce panning and zooming to canvas, rework properties gizmo into a drawer
This also changes how gizmos and some other
things are positioned and sized, to account for
new properties of the endless canvas.

While not utilized yet, this changeset also
includes the state manager for the undo/redo
system. It's a good idea to consider this from
the beginning, because the app is going to be
complex. The system was lifted from Bosca
Ceoil Blue, and should be good to go as is.

Smaller style adjustments have also been made.