Implement basic modifier key functionality for gizmos
A massive downside of the current
implementation is that changing modifiers while
dragging a gizmo doesn't retroactively apply
all previous changes with the new modifier logic.

This is done for simplicity, but this is not what
most users would expect to happen. So this
needs to be addressed at some point.