Rework property editors and improve their visuals
This was a necessary step to make the whole
system easier to use and extend. While it may
appear superficial, the rework allows for new
editor types to be added with a clean look
and clear function.

We also use scenes for these editors now,
so making layouts for them should be easier.
Additionally, a new trick is used to simplify
the theme setup. Something that we should
also extend to a few other widgets.