Implement basic text editing gizmo
This gizmo relies on a hidden text edit to do the
actual handling of text editing, so we can leverage
Godot's built-in feature and worry only about
high-level user interactions.

This can be a problem, because it's a nasty hack,
if I'm honest. But it beats trying to reimplement
all the peculiarities of text editing on all the
platforms. There is still quite a bit of data wrangling
trying to connect our gizmo, the hidden input,
and the text server. But it mostly works now!