WIP: Implement basic composite element and add support for it in the canvas drawer
There are still several features missing, most notably
the drawing logic is ad-hoc and doesn't use proper
nodes for child elements of composite elements. But
need to commit the progress so far.

Grouping is also done for all top-level elements in
a canvas. This is because we don't have means to
select multiple elements yet.