WIP: Rework some of the logic related to properties and states
This is a broken commit. The project needs a fresh
start/heavy refactoring at this point. This is to be
expected, as the current version has been developed
unevenly.

With the next commit I'll start again, using this
implementation for inspiration and reference, if
necessary. The focus should be on the breadth not
the depth. I.e. more systems need to be fully
developed and put in place, rather than more
variations and details should be polished and
implemented. The flow must be demonstratable
from beginning to the end, instead of pretty but
ultimately incomplete parts of the flow being worked
on.