WIP: Reimplement positioning and sizing for grouped/composite elements
This should be a more robust system than before.
We separate a request to redraw the proxy and a
request to reposition/resize it. The first one happens
as data changes, naturally. The second one happens
after the transform changes are propagated up the
tree.

When element has its spatial properties changed,
its data is updated immediately. However, we don't
tell the proxy to update its size and position just yet.
Instead, we inform owners of the element about the
changes. These owners, if they are composite
elements, update their own spatial properties in
turn. This repeats until we reach the canvas, or until
we reach the composite element that doesn't need
to change its position and size.

When this happens, we propagate back the signal
to actually notify about the transforms. This allows
us to update all spatial data on proxies starting from
the parent. You see, the data is in global
coordinates, but proxy controls are local to their
parent node. This allows us to reapply global data
without issues, since parents are updated first.