WIP: Implement states in composite elements
When element is parented to a composite element,
all composite's states are forced on that element.
When a state is added to a composite element, all
of its sub-elements are updated to also have that
state.

In the UI we will lock such states that are enforced
by the owner element. When the ownership ends,
states remain on elements, unless manually
removed.