WIP: Reimplement basics of property editors
This also simplifies variant states and removes the
extra data object per state. We are updating a data
object anyway, so we can just as well update the
state directly, without marking the property as an
override.

There are still some improvements to make later,
such as making sure that property editors are
updated for the selected state when default state
values change. But we'll figure that out when we
get there.