WIP: Implement a way to revert element/state properties
Variant state properties get reset to the default state
values. Default state properties get reset to the
default values defined in the script.

The implementation is very much ad-hoc, so it
likely be improved. We also don't have a decent way
to test if the current value can be reverted or not.
So no user indication, and the button is always
active. Something to improve for sure.