WIP: Rework property editors to be more self-contained
Also connect them to the owner element, not just
the state.

It's still a bit ugly, but at least we can use them as
types, without thinking about the underlying need
to instantiate associated scenes. The helper class
was basically an external way to do the same thing.
But, well, it was external, and the new static methods
are internal.