WIP: Optimize and improve transform notifications
The system is pretty straightforward:
-  Each element can have transform altering
changes, sometimes multiple in a short span of time.
- These changes are applied immediately, however
external observers must not react to them.
- Instead, there is a second notification that is
emitted with a delay at the end of the frame.
- So, elements signal transform_queued on every
related change, except there is safeguard and we
only ever allow it to be emitted once.
- Then the element owner tells it to flush its queued
transforms and report changes to external observers
via transform_changed.
- The element owner can be a composite element
or a canvas.
- Queued transforms are propagated upwards until
the canvas is reached, or one of composite elements
computes its values and sees no difference.
- Multiple elements can reach the canvas in the
same frame, so we defer until the end of frame to
process them.
- Calls that are stopped at a composite element are
not currently deferred.
- Then the transform_changed propagation starts
in the opposite direction, with last calculatations
being done and external observers updated.

With the current options available this minimizes the
number of calculations and updates needed. But it
can be optimized further, of course.